Remove the concept of 'channels' from non-core

+++ src/client.rs
@@ -3,5 +3,5 @@
 /// change (in the routegraph), renderer results, etc.
 pub trait Client {
     // TODO: use a multidimensional array type for buffer; remove num_ch param
-    fn audio_rendered(&mut self, _buffer: &[f32], _idx: u64, _num_ch: u8) {}
+    fn audio_rendered(&mut self, _buffer: &[f32], _idx: u64, _num_ch: u32) {}
 }
